"svn merge -r20031:20032 https://svn.openx.org/openx/branches/2.6/trunk ."
"svn merge -r20034:20035 https://svn.openx.org/openx/branches/2.6/trunk ."

+++ lib/OA/OperationInterval.php
@@ -136,7 +136,7 @@ class OA_OperationInterval
         $hours = $oDateCopy->getHour();
         $minutes = $oDateCopy->getMinute();
         $totalMinutes = ($days * 24 * 60) + ($hours * 60) + $minutes;
-        return(floor($totalMinutes / $operationInterval));
+        return (int) (floor($totalMinutes / $operationInterval));
     }
 
     /**
